fix data type of product and some authentication

+++ src/main/java/com/cafe/website/entity/Product.java
@@ -62,8 +62,8 @@ public class Product extends BaseEntity {
 
 	private String email;
 	private String facebook;
-	private Long latitude;
-	private Long longitude;
+	private Double latitude;
+	private Double longitude;
 	private String description;
 	private String location;
 
@@ -77,7 +77,7 @@ public class Product extends BaseEntity {
 	public Product(String name, String slug, String phone, List<Purpose> purposes, List<Area> areas, List<Kind> kinds,
 			List<Convenience> conveniences, List<Review> reviews, List<ProductSchedule> schedules, List<Menu> listMenus,
 			List<Image> listImages, User user, int priceMin, int priceMax, int outstanding, String email,
-			String facebook, Long latitude, Long longitude, String description, String location,
+			String facebook, Double latitude, Double longitude, String description, String location,
 			Boolean isWaitingDelete) {
 		super();
 		this.name = name;
@@ -240,19 +240,19 @@ public class Product extends BaseEntity {
 		this.facebook = facebook;
 	}
 
-	public Long getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(Long latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 
-	public Long getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(Long longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
